fix nnpack issues

+++ tiny_dnn/core/kernels/conv2d_op_nnpack.h
@@ -40,12 +40,13 @@ inline void conv2d_op_nnpack(const tensor_t &in_data,
   const float_t dy{0.0};  // params.in_padded.height_ - params.in.height_;
 
   // we'll assume that padding is symmetric
+  /* std::cout << "nnpack" << std::endl; */
 
   const nnp_padding padding = {
-    dy / 2),  // top
-    dx / 2),  // right
-    dy / 2),  // bottom
-    dx / 2)   // left
+    static_cast<size_t>(dy / 2),  // top
+    static_cast<size_t>(dx / 2),  // right
+    static_cast<size_t>(dy / 2),  // bottom
+    static_cast<size_t>(dx / 2)   // left
   };
 
   const nnp_size stride = {params.w_stride, params.h_stride};
